requirements.txt
Django==1.10.2
pkg-resources==0.0.0
plaid-python==1.4.0
requests==2.7.0
stripe==1.38.0
